# docs/openapi.yml
openapi: '3.1.0'
info:
  title: 'CareGuard API'
  version: '1.0.0'
  description: 'OpenAPI spec for CareGuard services: agent spending, pharmacy, bill audit, drug interactions, and payments.'
servers:
  - url: 'http://localhost:3000'
    description: 'Development server'
  - url: 'https://api.careguard.xyz'
    description: 'Production server'
security:
  - CaregiverBearerAuth: []
components:
  securitySchemes:
    CaregiverBearerAuth:
      type: 'http'
      scheme: 'bearer'
      bearerFormat: 'token'
      description: 'Bearer token used to authorize caregiver access to /agent/* endpoints. Set CAREGIVER_TOKEN in environment variables.'
    X402Auth:
      type: 'apiKey'
      name: 'X-Payment'
      in: 'header'
      description: 'x402 payment proof header for paid endpoints on Stellar.'
  schemas:
    Error:
      type: 'object'
      required:
        - 'error'
        - 'code'
      properties:
        error:
          type: 'string'
          description: 'A human-readable description of the error.'
        code:
          type: 'string'
          description: 'A stable, machine-readable uppercase SNAKE_CASE code identifying the specific type of error.'
        details:
          type: 'object'
          description: 'Structured metadata or field-level details specific to the error code.'
    X402PaymentChallenge:
      type: 'object'
      description: 'x402 payment challenge returned on 402 responses from paid routes. Shape is defined by the x402 protocol (see docs/setup/x402.md), not the shared Error schema — clients must branch on HTTP status 402, not on a `code` field, to detect this response.'
      properties:
        x402Version:
          type: 'integer'
        accepts:
          type: 'array'
          description: 'Accepted payment schemes, assets, and amounts for this route.'
          items:
            type: 'object'
        error:
          type: 'string'
          description: 'Human-readable reason payment is required or was rejected.'
...
paths:
...
  /bill/audit:
    post:
      summary: 'Audit medical bills'
      tags:
        - 'Bill Audit'
      security:
        - X402Auth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: 'object'
              properties:
                lineItems:
                  type: 'array'
                  items:
                    type: 'object'
                    properties:
                      cptCode:
                        type: 'string'
                      quantity:
                        type: 'integer'
                        minimum: 1
                        maximum: 999
                      chargedAmount:
                        type: 'number'
                        minimum: 0
                        maximum: 1000000
                      description:
                        type: 'string'
                        minLength: 1
                        maxLength: 80
                    required:
                      - 'cptCode'
                      - 'quantity'
                      - 'chargedAmount'
                      - 'description'
              required:
                - 'lineItems'
      responses:
        200:
          description: 'Audit results'
        402:
          description: 'Payment required. Response is an x402 payment challenge (not the shared Error schema): the body describes accepted payment schemes/amounts and an X-PAYMENT-related header identifies the facilitator. Retry the request with an X-PAYMENT header containing a valid payment proof for one of the accepted schemes.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/X402PaymentChallenge'
        413:
          description: 'Request body exceeds the configured size limit. `code` is BODY_TOO_LARGE; `details.limit` reports the configured maximum in bytes.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        429:
          description: 'Too many requests for this route''s rate-limit policy. `code` is RATE_LIMIT_EXCEEDED; back off and retry after the `Retry-After` header.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: 'Unhandled server error. `code` is SERVER_INTERNAL_ERROR.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...